+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.Web.UnitTests/Models/ApprenticeFilterModelTests/WhenGettingSortRouteData.cs
@@ -2,7 +2,7 @@
 using NUnit.Framework;
 using SFA.DAS.ProviderCommitments.Web.Models;
 
-namespace SFA.DAS.ProviderCommitments.Web.UnitTests.Models.ManageApprenticesFilterModelTests
+namespace SFA.DAS.ProviderCommitments.Web.UnitTests.Models.ApprenticeFilterModelTests
 {
     [TestFixture]
     public class WhenGettingSortRouteData
